+++ src/Candoumbe.Pipelines/Components/ICanSkipConfirmation.cs
@@ -1,5 +1,4 @@
 using Nuke.Common;
-using Nuke.Common.Tools.CorFlags;
 
 namespace Candoumbe.Pipelines.Components;
 /// <summary>
@@ -7,11 +6,10 @@ namespace Candoumbe.Pipelines.Components;
 /// </summary>/// <summary>
 /// Interface for components that allow skipping confirmation to end user
 /// </summary>
-public interface ICanSkipConfirmation
+public interface ICanSkipConfirmation : INukeBuild
 {
     /// <summary>
     /// Allow to skip asking confirmation to end user
     /// </summary>
-    [Parameter("Set to true to not ask any confirmation to the end user (default: false)")]
-    bool Quiet => false;
+    bool SkipConfirmation { get; }
 }